Keep Horizon autoscaler process counts integral
The autoscaler rounded worker allocations with ceil(), which returns a float,
then passed that value to ProcessPool::scale(int). Strict fixture typing
exposed the same TypeError that the real pool can encounter during scaling.
Convert the rounded count to int before applying the existing scaling limits.

Move FakePool into Fixtures and give it the real pool's native signatures so
existing autoscaler tests exercise this boundary. Remove redundant casts and
the duplicate promoted-property assignment from ProcessPool; retain the
existing scaling algorithm and public API.

Discovered while completing Laravel fixture normalization:
https://github.com/laravel/framework/pull/61315

Validated the existing autoscaler and supervisor tests, both scaling-direction
probes, full source/type analysis, formatting, and the full parallel suite.
